+++ compiler/src/iree/compiler/Codegen/LLVMGPU/TransformExtensions/LLVMGPUExtensions.cpp
@@ -11,6 +11,7 @@
 #include "iree/compiler/Codegen/Common/GPU/GPUVectorDistribution.h"
 #include "iree/compiler/Codegen/Common/GPU/Passes.h"
 #include "iree/compiler/Codegen/Dialect/Codegen/IR/IREECodegenAttrs.h"
+#include "iree/compiler/Codegen/Dialect/GPU/IR/GPUTileSwizzleUtils.h"
 #include "iree/compiler/Codegen/Dialect/GPU/IR/IREEGPUInterfaces.h"
 #include "iree/compiler/Codegen/LLVMGPU/Utils/LLVMGPUUtils.h"
 #include "iree/compiler/Codegen/Utils/GPUUtils.h"
@@ -1599,6 +1600,219 @@ createReadLayout(MLIRContext *ctx, const VectorExt::LayoutAttr &layout) {
   return VectorExt::LayoutAttr::get(ctx, perDimLayouts);
 }
 
+// Struct containing concrete MMA shape, type, and layout information.
+struct ConcreteMmaLayout {
+  GPU::OpaqueMmaLayout base;
+  VectorExt::PerDimLayoutAttr aMLayout;
+  VectorExt::PerDimLayoutAttr aKLayout;
+  VectorExt::PerDimLayoutAttr bKLayout;
+  VectorExt::PerDimLayoutAttr bNLayout;
+  VectorExt::PerDimLayoutAttr cMLayout;
+  VectorExt::PerDimLayoutAttr cNLayout;
+};
+
+static std::tuple<VectorExt::PerDimLayoutAttr, VectorExt::PerDimLayoutAttr>
+getPerDimLayoutAttrs(MLIRContext *context, TileSwizzle swizzle) {
+  // Step 1: obtain the swizzled tile shape, but keeping track of the source
+  // dimension indices.
+  struct SrcIndexAndSwizzleDim {
+    size_t srcIndex;
+    TileSwizzle::Dim dim;
+  };
+  SmallVector<SrcIndexAndSwizzleDim> swizzledShape;
+  for (auto [i, e] : llvm::enumerate(swizzle.expandShape)) {
+    for (TileSwizzle::Dim d : e) {
+      swizzledShape.push_back(SrcIndexAndSwizzleDim{i, d});
+    }
+  }
+  applyPermutationToVector(swizzledShape, swizzle.permutation);
+
+  // Step 2: collect the appropriate labels to use for the swizzled dims.
+  VectorExt::LayoutDimension internalLabels[] = {
+      VectorExt::LayoutDimension::VECTORZ, VectorExt::LayoutDimension::VECTORY,
+      VectorExt::LayoutDimension::VECTORX};
+  VectorExt::LayoutDimension crossThreadLabels[] = {
+      VectorExt::LayoutDimension::LANEZ, VectorExt::LayoutDimension::LANEY,
+      VectorExt::LayoutDimension::LANEX};
+  auto internalLabelIter = std::end(internalLabels);
+  auto crossThreadLabelIter = std::end(crossThreadLabels);
+  for (SrcIndexAndSwizzleDim d : swizzledShape) {
+    if (d.dim.kind == TileSwizzle::Dim::Kind::Internal) {
+      assert(internalLabelIter != std::begin(internalLabels));
+      --internalLabelIter;
+    } else if (d.dim.kind == TileSwizzle::Dim::Kind::CrossThread) {
+      assert(crossThreadLabelIter != std::begin(crossThreadLabels));
+      --crossThreadLabelIter;
+    } else {
+      assert(false && "unexpected dimension kind in intrinsic swizzle");
+    }
+  }
+
+  // Step 3: put together the result PerDimLayoutAttr'd for the two source dims.
+  SmallVector<VectorExt::LayoutDimensionAttr> labels[2];
+  SmallVector<int64_t> shape[2];
+  for (SrcIndexAndSwizzleDim d : swizzledShape) {
+    shape[d.srcIndex].push_back(d.dim.size);
+    auto &labelIterRef = (d.dim.kind == TileSwizzle::Dim::Kind::Internal)
+                             ? internalLabelIter
+                             : crossThreadLabelIter;
+    labels[d.srcIndex].push_back(VectorExt::LayoutDimensionAttr::get(
+        context, static_cast<VectorExt::LayoutDimension>(*labelIterRef++)));
+  }
+  return {VectorExt::PerDimLayoutAttr::get(context, labels[0], shape[0]),
+          VectorExt::PerDimLayoutAttr::get(context, labels[1], shape[1])};
+};
+
+static ConcreteMmaLayout getConcreteMMALayout(MLIRContext *context,
+                                              GPU::MMAIntrinsic intrinsic) {
+  auto opaque = GPU::getOpaqueMMALayout(context, intrinsic);
+  ConcreteMmaLayout concreteLayout;
+  concreteLayout.base = opaque;
+  auto lhsSwizzle = getIntrinsicSwizzle(intrinsic, GPU::MMAFragment::Lhs);
+  auto rhsSwizzle = getIntrinsicSwizzle(intrinsic, GPU::MMAFragment::Rhs);
+  auto accSwizzle = getIntrinsicSwizzle(intrinsic, GPU::MMAFragment::Acc);
+  std::tie(concreteLayout.aMLayout, concreteLayout.aKLayout) =
+      getPerDimLayoutAttrs(context, lhsSwizzle);
+  std::tie(concreteLayout.bNLayout, concreteLayout.bKLayout) =
+      getPerDimLayoutAttrs(context, rhsSwizzle);
+  std::tie(concreteLayout.cMLayout, concreteLayout.cNLayout) =
+      getPerDimLayoutAttrs(context, accSwizzle);
+  return concreteLayout;
+}
+
+static VectorExt::PerDimLayoutAttr
+getBatchedPerDimLayoutAttr(VectorExt::LayoutDimensionAttr batchDim,
+                           VectorExt::PerDimLayoutAttr baseLayout,
+                           int64_t problemSize, int64_t fragmentDimSize) {
+  assert(problemSize % fragmentDimSize == 0 &&
+         "invalid layout fragment for problem size");
+
+  SmallVector<VectorExt::LayoutDimensionAttr, 3> dimAttrs(
+      baseLayout.getLabels());
+  dimAttrs.insert(dimAttrs.begin(), batchDim);
+
+  SmallVector<int64_t, 3> shapes(baseLayout.getShapes());
+  shapes.insert(shapes.begin(), problemSize / fragmentDimSize);
+  auto layout = VectorExt::PerDimLayoutAttr::get(baseLayout.getContext(),
+                                                 dimAttrs, shapes);
+  return layout;
+}
+
+// Get the batched layout attributes for the given fragment layouts, indexing
+// map, and problem shape. The canonical fragment map is used to compare against
+// the problem map |indexingMap|. For example, for mma fragment B (RHS):
+//
+// indexingMap = affine_map<(d0, d1, d2) -> (d1, d2) # Transposed B
+// fragmentMap = affine_map<(d0, d1, d2) -> (d2, d1)
+// problemShape = [32, 64]
+// fragmentSize = [16, 8]
+// fragmentLayouts = [kLayout, nLayout]
+//
+// Gives batched layout
+//
+// Dim0 Layout = [BATCHX, nLayoutLabels], [8, nLayoutShape]
+// Dim1 Layout = [BATCHY, kLayoutLabels], [2, kLayoutShape]
+static VectorExt::LayoutAttr
+getBatchedLayoutAttr(AffineMap indexingMap, AffineMap fragmentMap,
+                     ArrayRef<int64_t> problemShape,
+                     ArrayRef<int64_t> fragmentSize,
+                     ArrayRef<VectorExt::PerDimLayoutAttr> fragmentLayouts) {
+  // Current distribution to MFMA operations does not support batched
+  // contractions so that is reflected here.
+  assert(indexingMap.getNumResults() == 2 &&
+         "invalid indexing map to non-batched simple contraction");
+
+  VectorExt::LayoutDimensionAttr batchX = VectorExt::LayoutDimensionAttr::get(
+      indexingMap.getContext(), VectorExt::LayoutDimension::BATCHX);
+  VectorExt::LayoutDimensionAttr batchY = VectorExt::LayoutDimensionAttr::get(
+      indexingMap.getContext(), VectorExt::LayoutDimension::BATCHY);
+
+  SmallVector<VectorExt::PerDimLayoutAttr, 2> perDimAttrs;
+  for (auto [expr, batchType] : llvm::zip_equal(
+           indexingMap.getResults(),
+           SmallVector<VectorExt::LayoutDimensionAttr, 2>{batchX, batchY})) {
+    auto maybeResultPosition = fragmentMap.getResultPosition(expr);
+    assert(maybeResultPosition && "fragment map and problem map mismatch");
+    int64_t idx = *maybeResultPosition;
+    perDimAttrs.push_back(getBatchedPerDimLayoutAttr(
+        batchType, fragmentLayouts[idx], problemShape[idx], fragmentSize[idx]));
+  }
+
+  return VectorExt::LayoutAttr::get(indexingMap.getContext(), perDimAttrs);
+}
+
+static FailureOr<std::tuple<VectorLayoutInterface, VectorLayoutInterface,
+                            VectorLayoutInterface>>
+getContractionLayout(vector::ContractionOp contract, ConcreteMmaLayout layout) {
+  MLIRContext *context = contract.getContext();
+  FailureOr<linalg::ContractionDimensions> maybeContractionDims =
+      linalg::inferContractionDims(contract.getIndexingMapsArray());
+  if (failed(maybeContractionDims)) {
+    return failure();
+  }
+  auto contractionDims = *maybeContractionDims;
+  // TODO: Relax this condition to strictly alignment requirements.
+  if (contractionDims.k.size() != 1 || contractionDims.m.size() != 1 ||
+      contractionDims.n.size() != 1) {
+    return failure();
+  }
+  // TODO: Support batched contractions.
+  if (contractionDims.batch.size() > 0) {
+    return failure();
+  }
+  unsigned mDim = contractionDims.m[0];
+  unsigned nDim = contractionDims.n[0];
+  unsigned kDim = contractionDims.k[0];
+
+  SmallVector<int64_t> iterationBounds;
+  contract.getIterationBounds(iterationBounds);
+
+  int64_t problemMSize = iterationBounds[mDim];
+  int64_t problemNSize = iterationBounds[nDim];
+  int64_t problemKSize = iterationBounds[kDim];
+
+  int64_t mSize = layout.base.mSize;
+  int64_t nSize = layout.base.nSize;
+  int64_t kSize = layout.base.kSize;
+
+  // The problem size currently must be strictly aligned to the size of the mma.
+  // This is expected to succeed assuming the correct [masked] vector size was
+  // set at strategy configuration time (for this mma).
+  if (problemMSize % mSize != 0 || problemNSize % nSize ||
+      problemKSize % kSize) {
+    return failure();
+  }
+
+  VectorExt::LayoutAttr aLayout = getBatchedLayoutAttr(
+      contract.getIndexingMapsArray()[0],
+      AffineMap::getMultiDimMapWithTargets(3, {mDim, kDim}, context),
+      {problemMSize, problemKSize}, {mSize, kSize},
+      {layout.aMLayout, layout.aKLayout});
+  VectorExt::LayoutAttr bLayout = getBatchedLayoutAttr(
+      contract.getIndexingMapsArray()[1],
+      AffineMap::getMultiDimMapWithTargets(3, {kDim, nDim}, context),
+      {problemKSize, problemNSize}, {kSize, nSize},
+      {layout.bKLayout, layout.bNLayout});
+  VectorExt::LayoutAttr cLayout = getBatchedLayoutAttr(
+      contract.getIndexingMapsArray()[2],
+      AffineMap::getMultiDimMapWithTargets(3, {mDim, nDim}, context),
+      {problemMSize, problemNSize}, {mSize, nSize},
+      {layout.cMLayout, layout.cNLayout});
+
+  return std::make_tuple<VectorLayoutInterface, VectorLayoutInterface,
+                         VectorLayoutInterface>(aLayout, bLayout, cLayout);
+}
+
+FailureOr<std::tuple<
+    VectorLayoutInterface, VectorLayoutInterface,
+    VectorLayoutInterface>> static getContractionLayout(GPU::MMAAttr mma,
+                                                        vector::ContractionOp
+                                                            contract) {
+  ConcreteMmaLayout layout = getConcreteMMALayout(
+      contract->getContext(), mma.getIntrinsic().getValue());
+  return getContractionLayout(contract, layout);
+}
+
 DiagnosedSilenceableFailure
 transform_dialect::SetContractionLayoutAttributes::apply(
     transform::TransformRewriter &rewriter,
@@ -1609,7 +1823,7 @@ transform_dialect::SetContractionLayoutAttributes::apply(
     return emitDefiniteFailure()
            << "invalid more than one attribute for contraction annotation";
   }
-  auto mmaType = llvm::dyn_cast<IREE::GPU::MmaInterfaceAttr>(typeList.front());
+  auto mmaType = llvm::dyn_cast<GPU::MMAAttr>(typeList.front());
   if (!mmaType) {
     return emitDefiniteFailure()
            << "invalid non-mma attribute for contraction annotation "
@@ -1623,7 +1837,7 @@ transform_dialect::SetContractionLayoutAttributes::apply(
              << "invalid non-contraction annotation " << payload;
     }
 
-    auto maybeLayouts = mmaType.getContractionLayout(contract);
+    auto maybeLayouts = getContractionLayout(mmaType, contract);
     if (failed(maybeLayouts)) {
       return emitDefiniteFailure()
              << "invalid opaque mma layout for annotation " << mmaType;
